fix(windows): native setjmp null frame, borrowed string retains, exact exit codes, Windows test lanes
Native EH on Windows called the CRT setjmp with an unset second argument.
The CRT stores that argument as the jump buffer's frame and longjmp runs an
SEH unwind toward any non-null frame, so a trap could unwind instead of
restoring the saved context. lowerNativeEh now takes a NativeSetjmpVariant
(Plain, DarwinMaskFree, WindowsNullFrame) and the Windows variant declares
setjmp(ptr, ptr) and passes an explicit null.

Borrowed runtime string results (Result.UnwrapStr, UnwrapErrStr,
Option.UnwrapStr) used directly as call arguments were released by the
frame after the owning Result or Option had already freed them. Both the
Zia and BASIC lowerers now retain a borrowed string at the call and queue
the usual statement-boundary release. ADR 0314 gains an amendment.

runExecutable returns std::optional<int> and reports the exact Windows exit
code through native_exit_code instead of a saturated value, so a native
program returning -6 matches the VM.

Windows build and test fixes:
- lroundf and strtoull resolve to ucrtbase/ucrtbased in the import planner
- D3D11 PerMaterial static assert matches the 528-byte temporal layout, the
  motion target uses the internal HDR16F format class, and the cut-budget
  override is forward-declared
- rt_postfx3d casts the LUT dimensions to int32_t
- isolated runtime tests pass /experimental:c11atomics per C source under
  Visual Studio generators
- native differential tests link each executable to a private build-tree
  path so concurrent VM/native and -O0/-O2 runs don't collide
- the x86-64 ABI probe pins SysV, unit tests avoid the `far` SDK macro,
  the Steam depot smoke avoids CRLF in its pack payload, the steam-linux
  JSON dry-run compares forward-slash paths, and g3d canvas tests get a
  60-second timeout

Tests: NativeEHLowering.WindowsSetjmpPassesNullFrame,
PlatformImportPlanners.WindowsRuntimeCrtConversionSymbolsResolveToUcrt, and
borrowed call-temporary cases in test_runtime_result_ownership.zia and
test_basic_runtime_result_ownership.bas.

+++ src/runtime/graphics/3d/render/rt_postfx3d.c
@@ -1239,8 +1239,8 @@ static int vgfx3d_postfx_fill_effect_snapshot(const rt_postfx3d *fx,
                 snapshot.color_lut_enabled = 1;
                 snapshot.color_lut_blend = e->p.color_lut.blend;
                 snapshot.color_lut_texels = lut->data;
-                snapshot.color_lut_width = lut->width;
-                snapshot.color_lut_height = lut->height;
+                snapshot.color_lut_width = (int32_t)lut->width;
+                snapshot.color_lut_height = (int32_t)lut->height;
                 snapshot.color_lut_revision = rt_pixels_generation(fx->lut_pixels);
             }
             break;
